Add files via upload
Technology has captured most of the fields in present and same goes for analyzing, modelling of data, with the advancement of the technology nowadays we can also predict the future results of our datasets using several machine learning algorithms and high level programming languages has made the job easier for example using several libraries in Python we can easily solve the problems of the huge datasets. The very first method of preserving the data was writing and it was Sumerians who invented writing around 3400 BC, imagine how difficult the world would be without the breakthrough of Sumerians with the only storage device “human brain” which for me is very unreliable storage for storing large data because human brain may have consciousness but its memory is very unreliable as compared to the writing or other form of modern storage devices. Keeping the following knowledge in mind, I became excited to solve the problem of “Loan approval dataset” which I found fascinating because of its importance in financial market and it is also helpful in analyzing which variable is highly correlated with the approval of loan and it is also fun to predict the future approval or disapproval of the loans. The dataset which I have used in this project is about the loan prediction, the goal of employing AI in this situation is to increase the speed and precision of the loan approval process. The manual assessment of applications used in traditional loan approval procedures can be time-consuming and error-prone. It may be able to automate some of the approval process and develop more precise predictions about which applications are most likely to be approved by utilising AI and machine learning techniques. To minimize bias and ensure fairness in the loan approval procedure is another justification for employing AI in this situation. Large datasets can be used to train AI models so they can find trends and make predictions based on data rather than opinions or biases. Overall, this dataset offers an exciting and timely challenge for the use of AI methods. It might be able to increase the efficiency and equity of lending by creating fair and accurate predictive models for loan acceptance, which would be advantageous to both lenders and borrowers.